server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pytesseract==0.3.13
Pillow==10.4.0
python-multipart==0.0.12
numpy==1.26.4
